src/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pycups==2.0.4
flake8==7.0.0
